--- packages/utils/camelize.js
@@ -1,4 +1,0 @@
-const camelizeRE = /-(\w)/g;
-export default function(str) {
-  return str.replace(camelizeRE, (_, c) => c ? c.toUpperCase() : '');
-}
