--- packages/utils/get.js
@@ -1,9 +1,0 @@
-export default function(object, path) {
-  const keys = path.split('.');
-  let result = object;
-
-  keys.forEach(key => {
-    result = result[key] || '';
-  });
-  return result;
-}
